refactor(tool): Move parser_factory into tool::parser
